refactoring measure.service attempt #1
